+++ frontend/src/components/Molecule/Chat/AvatarPopper.tsx
@@ -7,6 +7,7 @@ import { useNavigate, useParams } from 'react-router';
 import { API_URL } from '@/../config/backend';
 import GlobalContext from '@/context/GlobalContext';
 import { useError } from '@/context/ErrorContext';
+import { fetchDM } from '@/api/API';
 
 interface AvatarPopperProps {
   anchorEl: HTMLElement | null;
@@ -67,39 +68,7 @@ const AvatarPopper: FC<AvatarPopperProps> = ({ anchorEl, handleClose, target, sc
 
   function handleDMClick(id: number) {
     console.log(id + ' is DM');
-    async function fetchDM() {
-      const response = await fetch(API_URL + '/chat/dm', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          user_id: id,
-        }),
-      });
-      if (!response.ok) {
-        const error = await response.json();
-        handleError('Send DM', error.message);
-        return;
-      }
-      const ch = await response.json();
-      const newChannel = {
-        id: ch.channel_id,
-        type: ch.type,
-        title: ch.owner.nickname + '님과의 DM',
-        thumbnail: ch.owner.profile_url,
-        owner: {
-          id: ch.owner.user_id,
-          nickname: ch.owner.nickname,
-          profile: ch.owner.profile_url,
-        },
-      };
-      if (-1 === channels.findIndex((ch)=>(ch.id === newChannel.id)))
-        setChannels([newChannel, ...channels]);
-      navigate('/channels/' + ch.channel_id, { state: undefined });
-      console.log(response);
-    }
-    fetchDM();
+    fetchDM(id, navigate, handleError, channels, setChannels);
     handleClose();
   }
